fix: do not "uppercase" unit title of digital gauges
unit titles are automatically changed to uppercase. That is an issue when the unit is a greek letter. e.g. for the measure of a conductivity, the unit is µS/cm. Capitalizing it makes it MS/CM, which does not have the same meaning. Additionally, I don't understand the benefit of capitalizing this by default.